move mcp tool output types to dedicated file
