+++ front/lib/actions/mcp_internal_actions/utils.ts
@@ -1,4 +1,4 @@
-import type { MCPToolResult } from "@app/lib/actions/mcp_actions";
+import type { MCPToolResult } from "@app/lib/actions/mcp_internal_actions/output_schemas";
 
 export function makeMCPToolTextError(text: string): MCPToolResult {
   return {
