+++ front/lib/models/assistant/actions/mcp.ts
@@ -1,7 +1,7 @@
 import type { CreationOptional, ForeignKey, NonAttribute } from "sequelize";
 import { DataTypes } from "sequelize";
 
-import type { MCPToolResultContent } from "@app/lib/actions/mcp_actions";
+import type { MCPToolResultContent } from "@app/lib/actions/mcp_internal_actions/output_schemas";
 import { MCPServerViewModel } from "@app/lib/models/assistant/actions/mcp_server_view";
 import { AgentConfiguration } from "@app/lib/models/assistant/agent";
 import { AgentMessage } from "@app/lib/models/assistant/conversation";
